requirements.txt
pandas==1.3.5
psycopg2-binary==2.9.3
boto3==1.24.12
fsspec==0.7.4
s3fs==0.4.0
nba_api